OAK-11617: Provide oak-run commands to analyze and fix inconsistencies in the namespace registry
Some refactoring.
