+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/name/ReadOnlyNamespaceRegistry.java
@@ -25,28 +25,16 @@ import javax.jcr.NamespaceRegistry;
 import javax.jcr.RepositoryException;
 import javax.jcr.UnsupportedRepositoryOperationException;
 
-import org.apache.jackrabbit.oak.commons.collections.IterableUtils;
-import org.apache.jackrabbit.oak.commons.collections.SetUtils;
-import org.apache.jackrabbit.oak.commons.collections.StreamUtils;
-import org.apache.jackrabbit.oak.spi.nodetype.NodeTypeConstants;
 import org.apache.jackrabbit.oak.api.PropertyState;
 import org.apache.jackrabbit.oak.api.Root;
 import org.apache.jackrabbit.oak.api.Tree;
 import org.apache.jackrabbit.oak.spi.namespace.NamespaceConstants;
-import org.apache.jackrabbit.util.Text;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /**
  * Read-only namespace registry. Used mostly internally when access to the
@@ -140,196 +128,11 @@ public class ReadOnlyNamespaceRegistry
     }
 
     public boolean checkConsistency() throws IllegalStateException {
-        NamespaceRegistryModel model = NamespaceRegistryModel.create(namespaces);
-        boolean consistent = model.isConsistent();
-        if (!consistent) {
-            LOG.warn("Namespace registry is inconsistent. "
-                    + "Unregistered mapped prefixes: {}. "
-                    + "Unregistered mapped namespaces: {}. "
-                    + "Registered unmapped prefixes: {}. "
-                    + "Registered unmapped namespaces: {}.",
-                    model.getUnregisteredMappedPrefixes(),
-                    model.getUnregisteredMappedNamespaces(),
-                    model.getRegisteredUnmappedPrefixes(),
-                    model.getRegisteredUnmappedNamespaces());
-        }
-        return consistent;
+        return NamespaceRegistryModel.create(namespaces).isConsistent();
     }
 
     public NamespaceRegistryModel createNamespaceRegistryModel() {
         return NamespaceRegistryModel.create(namespaces);
     }
 
-    protected static final class NamespaceRegistryModel {
-        protected final Set<String> registeredPrefixes;
-        protected final Set<String> registeredNamespacesEncoded;
-        protected final Map<String, String> prefixToNamespaceMap;
-        protected final Map<String, String> namespaceToPrefixMap;
-
-        protected Set<String> mappedPrefixes;
-        protected Set<String> mappedNamespaces;
-        protected Set<String> mappedToPrefixes;
-        protected Set<String> mappedToNamespacesEncoded;
-        protected Set<String> allPrefixes;
-        protected Set<String> allNamespacesEncoded;
-        protected Set<String> consistentPrefixes;
-        protected Set<String> consistentNamespaces;
-        protected int registrySize;
-
-        private boolean consistent = false;
-        private boolean fixable = false;
-
-        private NamespaceRegistryModel(
-                Set<String> registeredPrefixes, Set<String> registeredNamespacesEncoded,
-                // prefixes to URIs
-                Map<String, String> prefixToNamespaceMap,
-                // encoded URIs to prefixes
-                Map<String, String> namespaceToPrefixMap) {
-            // ignore the empty namespace which is not mapped
-            this.registeredPrefixes = registeredPrefixes.stream().filter(s -> !(Objects.isNull(s) || s.isEmpty())).collect(Collectors.toSet());
-            this.registeredNamespacesEncoded = registeredNamespacesEncoded.stream().filter(s -> !(Objects.isNull(s) || s.isEmpty())).collect(Collectors.toSet());
-            this.prefixToNamespaceMap = prefixToNamespaceMap.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-            this.namespaceToPrefixMap = namespaceToPrefixMap.entrySet().stream().collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
-            init();
-        }
-
-        private void init() {
-            this.mappedPrefixes = prefixToNamespaceMap.keySet();
-            this.mappedNamespaces = namespaceToPrefixMap.keySet();
-            this.mappedToPrefixes = new HashSet<>(namespaceToPrefixMap.values());
-            this.mappedToNamespacesEncoded = prefixToNamespaceMap.values().stream().map(Namespaces::encodeUri).collect(Collectors.toSet());
-            allPrefixes = SetUtils.union(SetUtils.union(registeredPrefixes, mappedPrefixes), mappedToPrefixes);
-            allNamespacesEncoded = SetUtils.union(SetUtils.union(registeredNamespacesEncoded, mappedNamespaces), mappedToNamespacesEncoded);
-            registrySize = Math.max(allPrefixes.size(), allNamespacesEncoded.size());
-            consistentPrefixes = SetUtils.intersection(SetUtils.intersection(registeredPrefixes, mappedPrefixes), mappedToPrefixes);
-            consistentNamespaces = SetUtils.intersection(SetUtils.intersection(registeredNamespacesEncoded, mappedNamespaces), mappedToNamespacesEncoded);
-            consistent = consistentPrefixes.size() == consistentNamespaces.size()
-                    && consistentPrefixes.size() == allPrefixes.size();
-            if (consistent) {
-                fixable = true;
-            } else {
-                // everything needs to be contained in at least one of the bijective mappings
-                fixable = registrySize == SetUtils.union(mappedPrefixes, mappedToPrefixes).size()
-                        && registrySize == SetUtils.union(mappedNamespaces, mappedToNamespacesEncoded).size();
-            }
-        }
-
-        static NamespaceRegistryModel create(Tree namespaces) {
-            Tree nsdata = namespaces.getChild(REP_NSDATA);
-            Map<String, String> prefixToNamespaceMap = new HashMap<>();
-            Map<String, String> namespaceToPrefixMap = new HashMap<>();
-            for (PropertyState propertyState : namespaces.getProperties()) {
-                String prefix = propertyState.getName();
-                if (!prefix.equals(NodeTypeConstants.JCR_PRIMARYTYPE)) {
-                    prefixToNamespaceMap.put(prefix, propertyState.getValue(STRING));
-                }
-            }
-            for (PropertyState propertyState : nsdata.getProperties()) {
-                String encodedUri = propertyState.getName();
-                switch (encodedUri) {
-                    case REP_PREFIXES:
-                    case REP_URIS:
-                    case NodeTypeConstants.JCR_PRIMARYTYPE:
-                        break;
-                    default:
-                        namespaceToPrefixMap.put(encodedUri, propertyState.getValue(STRING));
-                }
-            }
-            Iterable<String> uris = nsdata.getProperty(REP_URIS).getValue(STRINGS);
-            NamespaceRegistryModel model = new NamespaceRegistryModel(
-                    new HashSet<>(Arrays.asList(IterableUtils.toArray(nsdata.getProperty(REP_PREFIXES).getValue(STRINGS), String.class))),
-                    StreamUtils.toStream(uris).map(Namespaces::encodeUri).collect(Collectors.toSet()),
-                    prefixToNamespaceMap, namespaceToPrefixMap);
-            return model;
-        }
-
-        NamespaceRegistryModel tryRegistryRepair() {
-            if (consistent) {
-                return this;
-            }
-            if (!fixable) {
-                return null;
-            }
-            HashSet<String> fixedRegisteredPrefixes = new HashSet<>();
-            HashMap<String, String> fixedPrefixToNamespaceMap = new HashMap<>();
-            for (String prefix : allPrefixes) {
-                fixedRegisteredPrefixes.add(prefix);
-                if (mappedPrefixes.contains(prefix)) {
-                    fixedPrefixToNamespaceMap.put(prefix, prefixToNamespaceMap.get(prefix));
-                } else {
-                    for (Map.Entry<String, String> entry : namespaceToPrefixMap.entrySet()) {
-                        if (entry.getValue().equals(prefix)) {
-                            fixedPrefixToNamespaceMap.put(prefix, Text.unescapeIllegalJcrChars(entry.getKey()));
-                            break;
-                        }
-                    }
-                }
-            }
-            HashSet<String> fixedRegisteredNamespacesEncoded = new HashSet<>();
-            HashMap<String, String> fixedNamespaceToPrefixMap = new HashMap<>();
-            for (String encodedNamespace : allNamespacesEncoded) {
-                fixedRegisteredNamespacesEncoded.add(encodedNamespace);
-                if (mappedNamespaces.contains(encodedNamespace)) {
-                    fixedNamespaceToPrefixMap.put(encodedNamespace, namespaceToPrefixMap.get(encodedNamespace));
-                } else {
-                    for (Map.Entry<String, String> entry : prefixToNamespaceMap.entrySet()) {
-                        if (Namespaces.encodeUri(entry.getValue()).equals(encodedNamespace)) {
-                            fixedNamespaceToPrefixMap.put(encodedNamespace, entry.getKey());
-                            break;
-                        }
-                    }
-                }
-            }
-           return new NamespaceRegistryModel(fixedRegisteredPrefixes, fixedRegisteredNamespacesEncoded,
-                   fixedPrefixToNamespaceMap, fixedNamespaceToPrefixMap);
-        }
-
-        boolean isConsistent() {
-            return consistent;
-        }
-
-        public boolean isFixable() {
-            return fixable;
-        }
-
-        Set<String> getUnregisteredMappedPrefixes() {
-            return SetUtils.difference(mappedPrefixes, registeredPrefixes);
-        }
-
-        Set<String> getRegisteredUnmappedPrefixes() {
-            return SetUtils.difference(registeredPrefixes, mappedPrefixes);
-        }
-
-        Set<String> getUnregisteredMappedNamespaces() {
-            return SetUtils.difference(mappedNamespaces, registeredNamespacesEncoded);
-        }
-
-        Set<String> getRegisteredUnmappedNamespaces() {
-            return SetUtils.difference(registeredNamespacesEncoded, mappedNamespaces);
-        }
-
-        Set<String> getRegisteredPrefixes() {
-            return registeredPrefixes;
-        }
-
-        Set<String> getRegisteredNamespacesEncoded() {
-            return registeredNamespacesEncoded;
-        }
-
-        Set<String> getMappedPrefixes() {
-            return mappedPrefixes;
-        }
-
-        Set<String> getMappedNamespaces() {
-            return mappedNamespaces;
-        }
-
-        Set<String> getAllPrefixes() {
-            return allPrefixes;
-        }
-
-        Set<String> getAllNamespaces() {
-            return allNamespacesEncoded;
-        }
-    }
 }
